move Room to js, reorganize files inside js, create action for setting speakers
