+++ packages/js/src/features/actions.ts
@@ -1,0 +1,5 @@
+import { createAction } from '@reduxjs/toolkit'
+
+export const audioSetSpeakerAction = createAction<string>(
+  'swJs/audioSetSpeakerAction'
+)
