+++ packages/js/src/index.ts
@@ -1,3 +1,5 @@
 export * as Video from './video'
 export * as WebRTC from './webrtc'
 export type { ClientEvents, UserOptions, Emitter } from '@signalwire/core'
+export { Room } from './Room'
+export * from './utils/interfaces'
