+++ packages/webrtc/src/utils/interfaces.ts
@@ -1,11 +1,3 @@
-import StrictEventEmitter from 'strict-event-emitter-types'
-import type {
-  RoomEventNames,
-  BaseConnectionState,
-  EventsHandlerMapping,
-} from '@signalwire/core'
-import type { Room } from '../Room'
-
 export interface ConnectionOptions {
   // TODO: Not used anymore but required for backend
   destinationNumber?: string
@@ -49,27 +41,4 @@ export interface ConnectionOptions {
   autoApplyMediaParams?: boolean
   rtcPeerConfig?: { [key: string]: any }
   iceGatheringTimeout?: number
-}
-
-type BaseConnectionEventsHandlerMapping = EventsHandlerMapping &
-  Record<BaseConnectionState, (params: Room) => void>
-
-export type RoomObjectEvents = {
-  [k in
-    | RoomEventNames
-    | BaseConnectionState]: BaseConnectionEventsHandlerMapping[k]
-}
-
-export type RoomObject = StrictEventEmitter<Room, RoomObjectEvents>
-
-export type CreateScreenShareObjectOptions = {
-  autoJoin?: boolean
-  audio?: MediaStreamConstraints['audio']
-  video?: MediaStreamConstraints['video']
-}
-
-export type CreateSecondSourceObjectOptions = {
-  autoJoin?: boolean
-  audio?: MediaStreamConstraints['audio']
-  video?: MediaStreamConstraints['video']
 }
